move exportbanner to its own file
